+++ tinper-bee/bee-datepicker/src/YearPicker.js
@@ -5,6 +5,7 @@
 import YearPanel from "./rc-calendar/year/YearPanel";
 import { KeyCode } from 'tinper-bee-core';
 import React, { Component } from "react";
+import ReactDOM from 'react-dom';
 import Picker from "./rc-calendar/Picker";
 import FormControl from "bee-form-control";
 import InputGroup from 'bee-input-group';
@@ -33,10 +34,16 @@ class YearPicker extends Component {
                 value: nextProps.value
             });
         }
-        this.setState({
-            renderIcon: nextProps.renderIcon,
-            open: nextProps.open
-        });
+        if ("open" in nextProps) {
+            this.setState({
+              open: nextProps.open 
+            });
+          }
+          if ("renderIcon" in nextProps) {
+            this.setState({
+              renderIcon: nextProps.renderIcon
+            });
+          }
 
     }
 
@@ -85,7 +92,7 @@ class YearPicker extends Component {
     handleChange = value => {
         const props = this.props;
         this.setState({ value });
-        props.onChange(value, (value && value.format(props.format)) || '');
+        props.onChange&&props.onChange(value, (value && value.format(props.format)) || '');
     }
     onMouseLeave = (e) => {
         this.setState({