+++ tinper-bee/bee-table/demo/demolist/Demo131.js
@@ -1,7 +1,7 @@
 /**
 *
 * @title 滚动加载
-* @description
+* @description 万行数据渲染
 */
 
 import React, { Component } from "react";