+++ tinper-bee/bee-datepicker/CHANGELOG.md
@@ -1,15 +1,45 @@
+<a name="2.0.23"></a>
+## [2.0.23](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.22...v2.0.23) (2019-04-12)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** yearpicker willReceiveProps 加判断 ([08e5974](https://github.com/tinper-bee/bee-datepicker/commit/08e5974))
+
+
+
+<a name="2.0.22"></a>
+## [2.0.22](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.21...v2.0.22) (2019-04-12)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** yearpicker onChange加判断 ([02edd44](https://github.com/tinper-bee/bee-datepicker/commit/02edd44))
+
+
+
+<a name="2.0.21"></a>
+## [2.0.21](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.20...v2.0.21) (2019-04-11)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** reactDOM未引入bug ([9654dfd](https://github.com/tinper-bee/bee-datepicker/commit/9654dfd))
+
+
+
 <a name="2.0.20"></a>
 ## [2.0.20](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.19...v2.0.20) (2019-04-03)
-
-
-
-<a name="2.0.19"></a>
-## [2.0.19](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.18...v2.0.19) (2019-04-03)
 
 
 ### Features
 
 * **bee-datepicker:** yearpicker 组件 快捷键 ([d85d125](https://github.com/tinper-bee/bee-datepicker/commit/d85d125))
+
+
+
+<a name="2.0.19"></a>
+## [2.0.19](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.18...v2.0.19) (2019-03-29)